Split Content.js into three part to prevent conflict.
